requirements.txt
magika==0.5.1
yara_python==4.2.3
yara_python_dex==1.0.4
